[PC-177] refactor: ValueItem 패키지 위치 변경
