+++ common/src/main/java/org/yapp/domain/value/ValueItem.java
@@ -1,4 +1,4 @@
-package org.yapp.domain.profile;
+package org.yapp.domain.value;
 
 import org.hibernate.annotations.Type;
